header container removed / new containers added.
I realized that I had one too many containers affecting the header section of the website. It was making it harder to align elements within the header so I removed it. However, I added a couple new containers: 'web-interact' and 'hero-main'. These containers make it easier to add properties and align within the second half of the header.