Further container implementation / text-alignment
container random-info and below were added to manipulate the section of the webpage with the boxes and subtext., as well as the quote part of the webpage. I also added a very useful property to the body and to the .testi container. It is a very fundemental, but very useful property. That is the text-align property. This serves as a reminder to DONT FORGET THE FUNDEMENTALS.