Use of inheritance and flex rules with containers
All of classes added here are used to line up the boxes and texts centerly. Using the rule that a container can only affect its immediate children items. I seperated the boxes from each other so they can still recieve the properties of box-1 class which line up the boxes and the texts laterally with each other, as well as create a gap in between the box and the subtext. The boxes class the encases the entire section contains properties that centers the content and creates gaps betweeen the content.