Call to action manipulation
The call to action section was created with the .section-four class. The height / width was set as well as the alignment of its items. The following section .cta-section sets the height / width of the blue cta section as well as the background color / color of text. The border-radius was also set to 6px which creates a rounded look on the edges of the cta section.